Localization prototype.  Lookup fails.
This is an earlier prototype which included an attempt at
loading the localized text.  However, the game stores the lookup
via some prefixes that are per type.

To continue, search for the Localization.Get() calls to find the
prefixes.

Note that the main branch is actually an ancestor of this commit.